requirements.txt
streamlit
openai
langchain
chromadb
pandas
numpy
scikit-learn
matplotlib
sentence-transformers
python-dotenv
requests
